seperating goldilocks math
